backend/requirements.txt
fastapi
uvicorn
python-jose[cryptograhpy]
passlib[bcrypt]
